Moved errors examples into examples dir
