+++ examples/errors/README.md
@@ -2,6 +2,6 @@
 
 The dataflows in this folder are here to check, for any possible kind of error, that it is easy to pinpoint where the erorr occurred.
 
-The `working.py` is the only dataflow that should correctly run. Don't mind if it does meaningless operations.
+The `working.py` is the only dataflow that should correctly run.
 
 Running each dataflow you should be able to fix the error easily by reading the error message.
