+++ examples/errors/setup_tracing_twice.py
